// src/app/features/entregas/models/entrega.model.ts
import { ArchivoDTO } from 'src/app/shared/models/common/archivo.model';
import { UsuarioOrganismoDTO } from 'src/app/shared/models/usuario/usuario-organismo.model';
import { EstadoEntrega } from '../enum/estado-entrega.enum';
import { TipoObservacion } from '../enum/tipo-observacion.enum';
import { TipoUnidad } from '../enum/tipo-unidad.enum';
import { IDescargoDTO } from './descargo.model';
import { IEntregableDTO } from './entregable.model';
import { ItemOrdenCompraDTO } from './item-orden-compra.model';

export interface IEntregaDTO {
  idEntrega?: number;
  entregable?: IEntregableDTO;
  itemOrdenCompra?: ItemOrdenCompraDTO;
  cantidad?: number;
  cantidadRecepcionAceptada?: number;
  cantidadRecepcionRechazada?: number;
  cantidadConformidadAceptada?: number;
  cantidadConformidadRechazada?: number;
  fechaEntrega?: string;
  fechaComprometida?: string;
  fechaRecepcion?: string;
  fechaConformidad?: string;
  estado?: EstadoEntrega;
  estadoRecepcion?: EstadoEntrega;
  responsable?: string;
  responsableRecepcion?: UsuarioOrganismoDTO;
  responsableConformidad?: UsuarioOrganismoDTO;
  documentos?: ArchivoDTO[];
  documentosRecepcion?: ArchivoDTO[];
  documentosConformidad?: ArchivoDTO[];
  motivoRecepcionRechazada?: string;
  motivoConformidadRechazada?: string;
  tipoUnidad?: TipoUnidad;
  puedeRecepcionEntregas?: boolean;
  puedeConformidadEntregas?: boolean;
  recepcionFueraFecha?: boolean;
  tipoObservacion?: TipoObservacion;
  observaciones?: string;
  seleccionado?: boolean;
  descargos?: IDescargoDTO[];
}